Adding kessler microphysics [WIP].

+++ Columbia/wrf_physics/module_mp_kessler.f95
@@ -20,39 +20,39 @@ CONTAINS
    !  taken from the COMMAS code - WCS 10 May 1999.
    !  converted from FORTRAN 77 to 90, tiled, WCS 10 May 1999.
 !----------------------------------------------------------------
-   REAL    , PARAMETER ::  c1 = .001 
-   REAL    , PARAMETER ::  c2 = .001 
-   REAL    , PARAMETER ::  c3 = 2.2 
-   REAL    , PARAMETER ::  c4 = .875 
-   REAL    , PARAMETER ::  fudge = 1.0 
-   REAL    , PARAMETER ::  mxfall = 10.0 
+   DOUBLE PRECISION    , PARAMETER ::  c1 = .001 
+   DOUBLE PRECISION    , PARAMETER ::  c2 = .001 
+   DOUBLE PRECISION    , PARAMETER ::  c3 = 2.2 
+   DOUBLE PRECISION    , PARAMETER ::  c4 = .875 
+   DOUBLE PRECISION    , PARAMETER ::  fudge = 1.0 
+   DOUBLE PRECISION    , PARAMETER ::  mxfall = 10.0 
 !----------------------------------------------------------------
    INTEGER,      INTENT(IN   )    :: ids,ide, jds,jde, kds,kde, &
                                      ims,ime, jms,jme, kms,kme, &
                                      its,ite, jts,jte, kts,kte
-   REAL   ,      INTENT(IN   )    :: xlv, cp
-   REAL   ,      INTENT(IN   )    :: EP2,SVP1,SVP2,SVP3,SVPT0
-   REAL   ,      INTENT(IN   )    :: rhowater
+   DOUBLE PRECISION   ,      INTENT(IN   )    :: xlv, cp
+   DOUBLE PRECISION   ,      INTENT(IN   )    :: EP2,SVP1,SVP2,SVP3,SVPT0
+   DOUBLE PRECISION   ,      INTENT(IN   )    :: rhowater
 
-   REAL, DIMENSION( ims:ime , kms:kme , jms:jme ),              &
+   DOUBLE PRECISION, DIMENSION( ims:ime , kms:kme , jms:jme ),              &
          INTENT(INOUT) ::                                       &
                                                             t , &
                                                             qv, &
                                                             qc, &
                                                             qr
 
-   REAL, DIMENSION( ims:ime , kms:kme , jms:jme ),              &
+   DOUBLE PRECISION, DIMENSION( ims:ime , kms:kme , jms:jme ),              &
          INTENT(IN   ) ::                                       &
                                                            rho, &
                                                            pii, &
                                                           dz8w 
 
-   REAL, DIMENSION( ims:ime , kms:kme , jms:jme ),              &
+   DOUBLE PRECISION, DIMENSION( ims:ime , kms:kme , jms:jme ),              &
          INTENT(IN   ) ::                                    z
 
-   REAL, INTENT(IN   ) :: dt_in
+   DOUBLE PRECISION, INTENT(IN   ) :: dt_in
 
-   REAL, DIMENSION( ims:ime , jms:jme ),                        &
+   DOUBLE PRECISION, DIMENSION( ims:ime , jms:jme ),                        &
          INTENT(INOUT) ::                               RAINNC, &
                                                        RAINNCV
 
@@ -60,17 +60,17 @@ CONTAINS
 
    ! local variables
 
-   REAL :: qrprod, ern, gam, rcgs, rcgsi
-   REAL, DIMENSION( its:ite , kts:kte, jts:jte ) ::     prod
-   REAL, DIMENSION(kts:kte) :: vt, prodk, vtden,rdzk,rhok,factor,rdzw
+   DOUBLE PRECISION :: qrprod, ern, gam, rcgs, rcgsi
+   DOUBLE PRECISION, DIMENSION( its:ite , kts:kte, jts:jte ) ::     prod
+   DOUBLE PRECISION, DIMENSION(kts:kte) :: vt, prodk, vtden,rdzk,rhok,factor,rdzw
    INTEGER :: i,j,k
    INTEGER :: nfall, n, nfall_new
-   REAL    :: qrr, pressure, temp, es, qvs, dz, dt
-   REAL    :: f5, dtfall, rdz, product
-   REAL    :: max_heating, max_condense, max_rain, maxqrp
-   REAL    :: vtmax, ernmax, crmax, factorn, time_sediment
-   REAL    :: qcr, factorr, ppt
-   REAL, PARAMETER :: max_cr_sedimentation = 0.75
+   DOUBLE PRECISION    :: qrr, pressure, temp, es, qvs, dz, dt
+   DOUBLE PRECISION    :: f5, dtfall, rdz, product
+   DOUBLE PRECISION    :: max_heating, max_condense, max_rain, maxqrp
+   DOUBLE PRECISION    :: vtmax, ernmax, crmax, factorn, time_sediment
+   DOUBLE PRECISION    :: qcr, factorr, ppt
+   DOUBLE PRECISION, PARAMETER :: max_cr_sedimentation = 0.75
 !----------------------------------------------------------------
 
    INTEGER :: imax, kmax
@@ -229,14 +229,13 @@ CONTAINS
        product = amax1(prod(i,k,j),-qc(i,k,j))
        t (i,k,j) = t(i,k,j) + gam*(product - ern)
        qv(i,k,j) = amax1(qv(i,k,j) - product + ern,0.)
-       qc(i,k,j) =       qc(i,k,j) + product
+       qc(i,k,j) = qc(i,k,j) + product 
        qr(i,k,j) = qr(i,k,j) - ern
 
      ENDDO
      ENDDO
 
   ENDDO  microphysics_outer_j_loop
-
   RETURN
 
   END SUBROUTINE kessler
